batch-processor: README section, TTL keep-alive, funder/token arg-order guard, preview_batch (#558, #559, #560, #561)
- #558: add a BatchTransferProcessor section to README (storage, signatures,
  validation order, error-code table, events) and link it from the component
  status table; sync docs/contract-errors.md, docs/security.md and
  docs/architecture.md
- #559: add contracts/batch-processor/src/ttl.rs and extend the instance TTL
  on a successful, non-empty process_batch, mirroring contracts/stream/src/ttl.rs;
  validation failures, the empty-batch no-op and preview_batch leave TTL alone
- #560: document the argument order on process_batch and enforce a real
  SEP-41 precondition (zero-address, G... wallet, balance probe) before
  funder.require_auth(), so a transposed funder/token pair returns InvalidToken
  instead of failing at host level
- #561: add read-only preview_batch sharing checks 1-4 with process_batch
  through the new validate_total helper (no funder, no token, no auth, no TTL)

Bump VERSION to 2 (behaviour changed) and regenerate .gas-snapshot.json:
process_batch_100 22,005,647 -> 22,066,828, new preview_batch_100 = 31,334.

Also fixes pre-existing breakage on main that blocked the workspace build,
clippy and CI: oracle parse error in fresh_prices check, governor DataKey
Copy derive, unused imports in stream and gas_snapshot, a rustfmt blank line,
and a stale token-vault event expectation (5 events, not 4).

+++ contracts/oracle/src/lib.rs
@@ -614,7 +614,7 @@ impl TwapOracle {
         // from fewer fresh feeders than `min_submitters` is not reliable
         // enough to return. `PriceStatus` exposes both counts so callers can
         // observe how far below quorum the set is.
-        if fresh_prices.len() as u32 < config.min_submitters {
+        if fresh_prices.len() < config.min_submitters {
             return Err(Error::InsufficientQuorum);
         }
